Upgarde to be used from node without babel
The compiled files under `./lib` has been added to the repo because
there is no plan to add the package to npm and to install it on a node
project the github link is used, so nothing like `.npmignore` or the
script `prepublish` can be used to install the package with the right
ECMA version. Also force the user to use babel is not an option.